migrate org customizations to class override mechanism.
